backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.12.0
plotly>=5.17.0
xarray>=2023.12.0
netcdf4>=1.6.0
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.2.0
pydantic>=2.5.0
pyarrow>=14.0.0